Add delete license test case

+++ license-service/src/main/java/com/gov/zw/service/LicenseServiceImpl.java
@@ -60,6 +60,11 @@ public class LicenseServiceImpl implements LicenseService {
 
     @Override
     public void removeLicense(License license) throws InvalidLicenseException {
-        this.licenseRepository.delete(license);
+        Optional<License> licenseOptional = Optional.ofNullable(license);
+        if (licenseOptional.isPresent()) {
+            this.licenseRepository.delete(license);
+        } else {
+            throw new InvalidLicenseException("The license is invalid!");
+        }
     }
 }
